move remaining applicable viewport.ts helpers to App.viewport.ts
